Add Vendas  Controller

+++ src/main/resources/db/migration/V001__cria-tabela-vendedores.sql
@@ -3,6 +3,6 @@ CREATE TABLE vendedores
     id                   SERIAL      NOT NULL,
     nome                 VARCHAR(60) NOT NULL,
     total_vendas         BIGINT      NOT NULL,
-    media_vendas_diarias BIGINT      NOT NULL,
+    media_vendas_diarias DECIMAL(10,2)      NOT NULL,
     PRIMARY KEY (id)
 );
